refactor parts_mgr
